requirements.txt
# Core Python dependencies for research project

# Knowledge Base
chromadb>=0.5.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Deep Learning & AI
torch>=2.0.0
tensorflow>=2.13.0

# Jupyter and Analysis
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Data Processing
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0


requests
beautifulsoup4
pandas
numpy
scikit-learn
matplotlib
seaborn
jupyter